Add files via upload
Lots of additions in this "update", mostly new sprites and corresponding objects.
Notice how to player is able to walk around, behind, and in front of in-game objects. This was accomplished using [depth = -bbox_bottom;] which makes the bottom of the collision mask (in this case only for the player) change as the player moves up and down. My understanding is that, in Game Maker, things with a higher number are drawn to the back and the lower the number the more forward they are drawn. Since the players depth variable is assigned -bbox_bottom, the player will always be drawn on top.

Other updates include sprites like tables, plants, a coffee mug, coffee grinder, bialetti mocha pot, picture frame, bed, and rug. What I am having trouble with at the moment is getting the aspect ratios right. The tutorial I am following is helping me understand better and I am sure I'll have a good grasp by the end of this project. 

Enjoy!